PopplerAnnotFreeText: re-use fallback font list from Annot class
We need to re-introduce a fallback font list because iOS
annotations with "/Helvetica" [1] cannot be resolved to
a proper font just by looking at the pdf structure.

However, instead of re-introducing a brand new fallback list,
we re-use the existing one in Annot::determineFallbackFont()
adding the "/Helvetica" case and making it public so
poppler-glib can use it.

Improves !1595