Mark updated indirect objects inside Resources as such
In AnnotWidget::generateFieldAppearance() we create an
Appearance stream with a Resources dict that is the
result of merging with the Form default Resources dict.

But in that merged Resources dict, if indirect object
(eg. a font) was updated with a new property obtained
from the Form default Resources dict, we were not marking
that indirect object as "updated", which would cause that
if the user "saved the pdf" that updated object won't be
copied over i.e. the old version would be used.

In the issue #1549 that caused the problematic pdf to
not save the indirect object (319 0) which was the /Helv font
with a merged Encoding dict from Form Default resources.

Fixes #1549