Avoid adding superflous popplerfont
PDF file of issue #1549 made Form::ensureFontsForAllCharacters()
to add a new popplerfont just to handle 'ñ' character, that is
not necessary and was caused by the fact that 'Resources' passed
to it had an incomplete /Font dictionary which lacked the /Encoding
dict thus provoking 'ñ' to not be decoded and so thinking it needed
a custom popplerfont for that.

'ñ' shows fine after typing it because the Appearance stream
created by AnnotWidget::generateFieldAppearance() merges the /Font
dictionary from Form default resources to the one in Field's
resources. So because the 'Encoding' dict is in the Form
default resources and gets merged to the Field's one and this
combined /Font is the one used in the AP Resources that's why
it displays fine the 'ñ' character.

On the contrary, inside FormFieldText::setContentCopy() we
call Form::ensureFontsForAllCharacters() and currently pass it
the different existant Resources (Field, Form and AP) and the
looked up font is found in the Field's one, but that one lacks
the /Encoding dict. So the solution is to do same as generateFieldAppearance()
and merge Field and Form Resources dict prior to pass it to
ensureFontsForAllCharacters() so that when looking up the font
it finds it with the /Encoding dict in it.

As part of that solution we moved the recursiveMergeDicts()
function from Annot.cc to a static method in Dict class so
we can call it from FormFieldText::setContentCopy() in Form.cc

Part of #1549