Form: Do not iterate the document stream concurrently
When using validateSignatureAsync from multiple threads we end up
crashing because each thread may want to go through the document thread
messing up with it, changing its pointer and thus corrupting the memory.

So, protect the whole streaming phase with a mutex and then continue
asynchronously again.

Ideally each time doc->getBaseStream() the buffer should be locked to
that thread, but that implies a bigger change, so this is enough for now
given that validateSignatureAsync is the only API we want to actually
ensure to work concurrently since it may have a relevant use.

Closes: #1585