NSS: Add ability to use a password callback with data and destroy function
In order to unlock the NSS database we expose a function that allows to
set a function callback when the password is required.

This works well for C++ backends because of the existence of lambda
functions that allow to natively and safely pass data to it, but it's
not great for GLib and C bindings in general because this doesn't allow
to pass any data to the callback for better context.

And they have to re-implement this locally with static values that are
far from clean and don't allow to have automatic memory management of
the callback data.

So, add a new method that takes a function (with more parameters), the
function data and the function data destroyer.

Now, every time the password callback is set, this is bound to some
callback data too.

And such data is destroyed on password handler changes or when nss is
shut down.