introduce object::getNameString and apply it
In many cases, we take a object name char pointer and stuff it directly
into a Goo or std::string. Let that conversion happen in Object instead
of every location it is used.
This helps reduce the usage of the char pointer constructor in GooString
that is one of the slightly surprising things of GooString, to help
being able to kill it off at some point.

Also, Icon's are simple strings and can just as well be std::string, and
the same for text encoding